JNI: Added 64 bit support (hopefully working...)
git-svn-id: https://svn-i2.informatik.rwth-aachen.de/repos/libalf/trunk@495 ee1db7db-6c55-0410-a809-e4d5348c71a5

+++ JNI/java/de/libalf/jni/JNIAlgorithm.java
@@ -23,28 +23,28 @@ public abstract class JNIAlgorithm extends LibALFObject implements LearningAlgor
 		add_counterexample(counterexample, this.pointer);
 	}
 
-	private native void add_counterexample(int[] counterexample, int pointer);
+	private native void add_counterexample(int[] counterexample, long pointer);
 
 	@Override
 	public BasicAutomaton advance() {
 		return advance(this.pointer);
 	}
 
-	private native BasicAutomaton advance(int pointer);
+	private native BasicAutomaton advance(long pointer);
 
 	@Override
 	public boolean conjecture_ready() {
 		return conjecture_ready(this.pointer);
 	}
 
-	private native boolean conjecture_ready(int pointer);
+	private native boolean conjecture_ready(long pointer);
 
 	@Override
 	public int get_alphabet_size() {
 		return get_alphabet_size(this.pointer);
 	}
 
-	private native int get_alphabet_size(int pointer);
+	private native int get_alphabet_size(long pointer);
 
 	@Override
 	public void increase_alphabet_size(int newSize) {
@@ -52,7 +52,7 @@ public abstract class JNIAlgorithm extends LibALFObject implements LearningAlgor
 
 	}
 
-	private native void increase_alphabet_size(int newSize, int pointer);
+	private native void increase_alphabet_size(int newSize, long pointer);
 
 	@Override
 	public void set_alphabet_size(int alphabetSize) {
@@ -60,7 +60,7 @@ public abstract class JNIAlgorithm extends LibALFObject implements LearningAlgor
 
 	}
 
-	private native void set_alphabet_size(int alphabetSize, int pointer);
+	private native void set_alphabet_size(int alphabetSize, long pointer);
 
 	@Override
 	public void set_knowledge_source(Knowledgebase base) {
@@ -70,43 +70,43 @@ public abstract class JNIAlgorithm extends LibALFObject implements LearningAlgor
 			set_knowledge_source(base.getPointer(), this.pointer);
 	}
 
-	private native void set_knowledge_source(int knowledgebase_pointer,
-			int pointer);
+	private native void set_knowledge_source(long knowledgebase_pointer,
+			long pointer);
 
-	private native void set_knowledge_source_NULL(int pointer);
+	private native void set_knowledge_source_NULL(long pointer);
 
 	@Override
 	public boolean sync_to_knowledgebase() {
 		return sync_to_knowledgebase(this.pointer);
 	}
 
-	private native boolean sync_to_knowledgebase(int pointer);
+	private native boolean sync_to_knowledgebase(long pointer);
 
 	@Override
 	public boolean supports_sync() {
 		return supports_sync(this.pointer);
 	}
 
-	private native boolean supports_sync(int pointer);
+	private native boolean supports_sync(long pointer);
 
 	@Override
 	public int[] serialize() {
 		return serialize(this.pointer);
 	}
 
-	private native int[] serialize(int pointer);
+	private native int[] serialize(long pointer);
 	
 	@Override
 	public boolean deserialize(int[] serialization) {
 		return deserialize(serialization, this.pointer);
 	}
 
-	private native boolean deserialize(int[] serialization, int pointer);
+	private native boolean deserialize(int[] serialization, long pointer);
 
 	@Override	
 	public String toString() {
 		return tostring(this.pointer);
 	}
 
-	private native String tostring(int pointer);
+	private native String tostring(long pointer);
 }
